code_runner/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiopyston==1.2.1
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.1.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
pika==1.3.2
propcache==0.3.0
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-dotenv==1.0.1
typing_extensions==4.12.2
yarl==1.18.3
